:truck: move utils to their own directory
